feat : aop 적용 및, 사장 확인 로직 추가

+++ src/main/java/hairSalonReservation/sideProject/domain/serviceMenu/service/ServiceMenuService.java
@@ -1,5 +1,6 @@
 package hairSalonReservation.sideProject.domain.serviceMenu.service;
 
+import hairSalonReservation.sideProject.common.annotation.CheckRole;
 import hairSalonReservation.sideProject.domain.serviceMenu.dto.request.CreateServiceMenuRequest;
 import hairSalonReservation.sideProject.domain.serviceMenu.dto.request.UpdateServiceMenuRequest;
 import hairSalonReservation.sideProject.domain.serviceMenu.dto.response.ServiceMenuResponse;
@@ -9,11 +10,11 @@ import hairSalonReservation.sideProject.domain.serviceMenu.repository.ServiceMen
 import hairSalonReservation.sideProject.domain.serviceMenu.repository.ServiceMenuRepository;
 import hairSalonReservation.sideProject.domain.serviceMenu.repository.ServiceMenuRepositoryCustomImpl;
 import hairSalonReservation.sideProject.global.exception.ErrorCode;
+import hairSalonReservation.sideProject.global.exception.ForbiddenException;
 import hairSalonReservation.sideProject.global.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 
 @Service
@@ -25,6 +26,7 @@ public class ServiceMenuService {
     private final ServiceMenuRepositoryCustomImpl serviceMenuRepositoryCustom;
     private final ServiceMenuCategoryMapperRepository serviceMenuCategoryMapperRepository;
 
+    @CheckRole("OWNER")
     @Transactional
     public ServiceMenuResponse createServiceMenu(Long serviceCategoryMapperId, CreateServiceMenuRequest request){
 
@@ -48,11 +50,15 @@ public class ServiceMenuService {
                 .stream().map(ServiceMenuResponse::from).toList();
     }
 
+    @CheckRole("OWNER")
     @Transactional
-    public ServiceMenuResponse updateServiceMenu(Long serviceMenuId, UpdateServiceMenuRequest request){
+    public ServiceMenuResponse updateServiceMenu(Long userId, Long serviceMenuId, UpdateServiceMenuRequest request){
 
         ServiceMenu menu = serviceMenuRepository.findById(serviceMenuId)
                 .orElseThrow(() -> new NotFoundException(ErrorCode.SERVICE_MENU_NOT_FOUND));
+
+        Long shopOwnerId = menu.getServiceMenuCategoryMapper().getDesigner().getShop().getId();
+        if(!userId.equals(shopOwnerId)){throw new ForbiddenException(ErrorCode.FORBIDDEN);}
 
         menu.update(
                 request.name(),
@@ -63,11 +69,15 @@ public class ServiceMenuService {
         return ServiceMenuResponse.from(menu);
     }
 
+    @CheckRole("OWNER")
     @Transactional
-    public void deleteServiceMenu(Long serviceMenuId){
+    public void deleteServiceMenu(Long userId, Long serviceMenuId){
 
         ServiceMenu menu = serviceMenuRepository.findById(serviceMenuId)
                 .orElseThrow(() -> new NotFoundException(ErrorCode.SERVICE_MENU_NOT_FOUND));
+
+        Long shopOwnerId = menu.getServiceMenuCategoryMapper().getDesigner().getShop().getId();
+        if(!userId.equals(shopOwnerId)){throw new ForbiddenException(ErrorCode.FORBIDDEN);}
 
         menu.delete();
     }
